+++ src/chunks/track.ts
@@ -1,5 +1,5 @@
 import {Constants} from '../constants';
-import {ControllerChangeEvent} from '../meta-events/controller-change-event';
+import {ControllerChangeEvent} from '../midi-events/controller-change-event';
 import {CopyrightEvent} from '../meta-events/copyright-event';
 import {CuePointEvent} from '../meta-events/cue-point-event';
 import {EndTrackEvent} from '../meta-events/end-track-event';
@@ -10,7 +10,7 @@ import {MarkerEvent} from '../meta-events/marker-event';
 import {NoteEvent} from '../midi-events/note-event';
 import {NoteOnEvent} from '../midi-events/note-on-event';
 import {NoteOffEvent} from '../midi-events/note-off-event';
-import {PitchBendEvent} from '../meta-events/pitch-bend-event';
+import {PitchBendEvent} from '../midi-events/pitch-bend-event';
 import {TempoEvent} from '../meta-events/tempo-event';
 import {TextEvent} from '../meta-events/text-event';
 import {TimeSignatureEvent} from '../meta-events/time-signature-event';
