+++ src/midi-events/controller-change-event.ts
@@ -6,7 +6,8 @@ import {Utils} from '../utils.js';
  * @param {object} fields {controllerNumber: integer, controllerValue: integer, delta: integer}
  * @return {ControllerChangeEvent}
  */
-class ControllerChangeEvent implements MetaEvent, AbstractEvent {
+class ControllerChangeEvent implements MidiEvent, AbstractEvent {
+	channel: number;
 	data: number[];
 	delta: number;
 	type: string;
