+++ src/midi-events/midi-event.ts
@@ -1,5 +1,7 @@
 interface MidiEvent {
     channel: number;
+    data: number[];
     delta: number;
+    status: number;
     type: string;
 }
