+++ src/midi-events/note-event.ts
@@ -7,7 +7,7 @@ import {Utils} from '../utils';
  * @param {object} fields - {pitch: '[C4]', duration: '4', wait: '4', velocity: 1-100}
  * @return {NoteEvent}
  */
-class NoteEvent implements MidiEvent {
+class NoteEvent implements AbstractEvent {
 	data: number[];
 	delta: number;
 	events: MidiEvent[];
