+++ src/midi-events/note-on-event.ts
@@ -9,7 +9,10 @@ class NoteOnEvent implements MidiEvent, AbstractEvent {
 	channel: number;
 	data: number[];
 	delta: number;
+	status: number;
 	type: string;
+	pitch: string|string[];
+	velocity: number;
 
 	constructor(fields) {
 		// Set default fields
@@ -30,6 +33,7 @@ class NoteOnEvent implements MidiEvent, AbstractEvent {
 		this.tick 		= null;
 		this.delta 		= null;
 		this.data 		= fields.data;
+		this.status = 0x90;
 	}
 
 	/**
@@ -58,21 +62,13 @@ class NoteOnEvent implements MidiEvent, AbstractEvent {
 
 		this.data = Utils.numberToVariableLength(this.deltaWithPrecisionCorrection)
 					.concat(
-							this.getStatusByte(),
+						this.status | this.channel - 1,
 							Utils.getPitch(this.pitch, options.middleC),
 							Utils.convertVelocity(this.velocity)
 					);
 
 		return this;
 	}
-
-	/**
-	 * Gets the note on status code based on the selected channel. 0x9{0-F}
-	 * Note on at channel 0 is 0x90 (144)
-	 * 0 = Ch 1
-	 * @return {number}
-	 */
-	getStatusByte() {return 144 + this.channel - 1}
 }
 
 export {NoteOnEvent};
